# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HiraNXT Jewellery ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchases:
    get:
      operationId: listPurchases
      tags: [purchases]
      summary: List all purchases
      responses:
        "200":
          description: Purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Create a purchase entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "201":
          description: Created purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}:
    get:
      operationId: getPurchase
      tags: [purchases]
      summary: Get a purchase
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
    patch:
      operationId: updatePurchase
      tags: [purchases]
      summary: Update a purchase
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseUpdate"
      responses:
        "200":
          description: Updated purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}/cancel:
    post:
      operationId: cancelPurchase
      tags: [purchases]
      summary: Cancel a purchase (reverses its journal entry; only before any follow-up payment)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"

  # ---------- SUPPLIERS ----------
...
components:
  schemas:
...
    Purchase:
      type: object
      required: [id, supplierId, supplierName, metalType, purity, grossWeight, netWeight, fineWeight, ratePerGram, totalAmount, paidAmount, balanceAmount, paymentMode, taxableValue, gstRate, gstAmount, hsnCode, invoiceNumber, purchaseDate, cancelledAt, createdAt]
      properties:
        id:
          type: integer
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: string
        metalType:
          type: string
        purity:
          type: string
        grossWeight:
          type: number
        netWeight:
          type: number
        fineWeight:
          type: number
        ratePerGram:
          type: number
        totalAmount:
          type: number
        paidAmount:
          type: number
        balanceAmount:
          type: number
        paymentMode:
          type: string
        bankAccountId:
          type: ["integer", "null"]
        taxableValue:
          type: ["number", "null"]
        gstRate:
          type: ["number", "null"]
        gstAmount:
          type: ["number", "null"]
        hsnCode:
          type: ["string", "null"]
        invoiceNumber:
          type: string
        purchaseDate:
          type: string
        notes:
          type: ["string", "null"]
        cancelledAt:
          type: ["string", "null"]
        createdAt:
          type: string
    PurchaseInput:
      type: object
      required: [supplierName, metalType, purity, grossWeight, netWeight, fineWeight, ratePerGram, totalAmount, purchaseDate]
      properties:
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: string
        metalType:
          type: string
        purity:
          type: string
        grossWeight:
          type: number
        netWeight:
          type: number
        fineWeight:
          type: number
        ratePerGram:
          type: number
        totalAmount:
          type: number
        paidAmount:
          type: number
        paymentMode:
          type: string
        bankAccountId:
          type: ["integer", "null"]
        gstRate:
          type: number
        hsnCode:
          type: string
        purchaseDate:
          type: string
        notes:
          type: ["string", "null"]
    PurchaseUpdate:
      type: object
      properties:
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: string
        metalType:
          type: string
        purity:
          type: string
        grossWeight:
          type: number
        netWeight:
          type: number
        fineWeight:
          type: number
        ratePerGram:
          type: number
        totalAmount:
          type: number
        paidAmount:
          type: number
        paymentMode:
          type: string
        bankAccountId:
          type: ["integer", "null"]
        gstRate:
          type: ["number", "null"]
        hsnCode:
          type: string
        purchaseDate:
          type: string
        notes:
          type: ["string", "null"]
...