# lib/api-zod/src/generated/types/purchase.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HiraNXT Jewellery ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface Purchase {
  id: number;
  /** @nullable */
  supplierId: number | null;
  supplierName: string;
  metalType: string;
  purity: string;
  grossWeight: number;
  netWeight: number;
  fineWeight: number;
  ratePerGram: number;
  totalAmount: number;
  paidAmount: number;
  balanceAmount: number;
  paymentMode: string;
  /** @nullable */
  bankAccountId?: number | null;
  /** @nullable */
  taxableValue: number | null;
  /** @nullable */
  gstRate: number | null;
  /** @nullable */
  gstAmount: number | null;
  /** @nullable */
  hsnCode: string | null;
  invoiceNumber: string;
  purchaseDate: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  cancelledAt: string | null;
  createdAt: string;
}